# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Dev4AI Product API
  description: Swagger documentation for the Express product management backend.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful
...
    Company:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "2b0d7a0d-fb90-4ff3-a4b7-0d5f5c66f8d8"
        name:
          type: string
          example: "Acme Inc"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Squad:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "8e2b6d7f-0af2-4d19-bf31-f886f80d29d1"
        companyId:
          type: string
          format: uuid
          example: "2b0d7a0d-fb90-4ff3-a4b7-0d5f5c66f8d8"
        name:
          type: string
          example: "Platform"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Sprint:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "c231a116-e5ef-49aa-b0cc-7aeb7dd69f79"
        squadId:
          type: string
          format: uuid
          example: "8e2b6d7f-0af2-4d19-bf31-f886f80d29d1"
        name:
          type: string
          example: "Sprint 1"
        status:
          type: string
          enum: [PLANNED, ACTIVE, COMPLETED]
          example: PLANNED
        startDate:
          type: string
          format: date-time
          example: "2026-08-20T00:00:00.000Z"
        endDate:
          type: string
          format: date-time
          example: "2026-08-30T00:00:00.000Z"
    Ticket:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "7d4831fd-2dca-41d6-8914-5d9d9d4d3bcb"
        squadId:
          type: string
          format: uuid
          example: "8e2b6d7f-0af2-4d19-bf31-f886f80d29d1"
        sprintId:
          type: string
          format: uuid
          example: "c231a116-e5ef-49aa-b0cc-7aeb7dd69f79"
        title:
          type: string
          example: "Create dashboard filters"
        description:
          type: string
          example: "Add filters for team and sprint selection"
        status:
          type: string
          enum: [TODO, IN_PROGRESS, DONE]
          example: TODO
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, URGENT]
          example: HIGH
        createdBy:
          type: string
          format: uuid
          example: "7d1fbf7a-0d7d-48d5-a63d-c14d8d1a3bcd"
        assignedTo:
          type: string
          format: uuid
          nullable: true
          example: "e897af10-e6a0-4293-a57d-79e35d26152c"
...
    CompanyCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "Acme Inc"
...
    SquadCreateRequest:
      type: object
      required: [companyId, name]
      properties:
        companyId:
          type: string
          format: uuid
          example: "2b0d7a0d-fb90-4ff3-a4b7-0d5f5c66f8d8"
        name:
          type: string
          example: "Platform"
...
    SquadMember:
      type: object
      properties:
        squadId:
          type: string
          format: uuid
          example: "8e2b6d7f-0af2-4d19-bf31-f886f80d29d1"
        userId:
          type: string
          format: uuid
          example: "7d1fbf7a-0d7d-48d5-a63d-c14d8d1a3bcd"
        joinedAt:
          type: string
          format: date-time
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            email:
              type: string
              format: email
            role:
              type: string
...
    SquadMemberCreateRequest:
      type: object
      required: [squadId, userId]
      properties:
        squadId:
          type: string
          format: uuid
          example: "8e2b6d7f-0af2-4d19-bf31-f886f80d29d1"
        userId:
          type: string
          format: uuid
          example: "7d1fbf7a-0d7d-48d5-a63d-c14d8d1a3bcd"
    SprintCreateRequest:
      type: object
      required: [squadId, name, startDate, endDate]
      properties:
        squadId:
          type: string
          format: uuid
          example: "8e2b6d7f-0af2-4d19-bf31-f886f80d29d1"
        name:
          type: string
          example: "Sprint 1"
        startDate:
          type: string
          format: date-time
          example: "2026-08-20T00:00:00.000Z"
        endDate:
          type: string
          format: date-time
          example: "2026-08-30T00:00:00.000Z"
...
    TicketCreateRequest:
      type: object
      required: [squadId, sprintId, title]
      properties:
        squadId:
          type: string
          format: uuid
          example: "8e2b6d7f-0af2-4d19-bf31-f886f80d29d1"
        sprintId:
          type: string
          format: uuid
          example: "c231a116-e5ef-49aa-b0cc-7aeb7dd69f79"
        title:
          type: string
          example: "Create dashboard filters"
        description:
          type: string
          example: "Add filters for team and sprint selection"
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, URGENT]
          example: HIGH
          description: "Defaults to MEDIUM if not provided"
        assignedTo:
          type: string
          format: uuid
          nullable: true
          example: "e897af10-e6a0-4293-a57d-79e35d26152c"
          description: "Must be a member of the squad. createdBy is derived from the authenticated JWT — do NOT send it in the request body."
...
paths:
...
  /auth/github:
    get:
      tags: [Auth]
      summary: Initiate GitHub OAuth login
      description: Generates a cryptographically secure CSRF state token and PKCE code challenge, sets an HttpOnly verification cookie, and redirects the browser to GitHub's authorization consent screen.
      responses:
        '302':
          description: Redirects to GitHub authorization endpoint
          headers:
            Location:
              schema:
                type: string
                example: "https://github.com/login/oauth/authorize?client_id=...&scope=read:user+user:email&state=..."
            Set-Cookie:
              schema:
                type: string
                example: "devai_oauth_nonce=...; Path=/auth/github; HttpOnly; SameSite=Lax"
        '200':
          description: Returns authorization URL if requested with Accept application/json header
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          authUrl:
                            type: string
                          state:
                            type: string
...
  /companies:
    post:
      tags: [Company]
      summary: Create a company
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyCreateRequest'
      responses:
        '201':
          description: Company created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Company'
    get:
      tags: [Company]
      summary: Get all companies
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Companies fetched successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Company'
...
  /squads:
    post:
      tags: [Squad]
      summary: Create a squad
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SquadCreateRequest'
      responses:
        '201':
          description: Squad created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Squad'
...
  /squad-members:
    post:
      tags: [Squad Member]
      summary: Add a member to a squad via squad-members
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SquadMemberCreateRequest'
      responses:
        '201':
          description: Member added successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SquadMember'
...
  /sprints:
    post:
      tags: [Sprint]
      summary: Create a sprint
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SprintCreateRequest'
      responses:
        '201':
          description: Sprint created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Sprint'
...
  /tickets:
    post:
      tags: [Ticket]
      summary: Create a ticket
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCreateRequest'
      responses:
        '201':
          description: Ticket created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Ticket'
...